Cleaned up the organization of files
