+++ UnfinishedNotebooks/pingGraphQL.py
@@ -1,0 +1,36 @@
+import json
+import pandas as pd
+import requests
+import matplotlib.pyplot as plt
+
+
+def get_leaderboard() -> pd.DataFrame:
+    """
+        Ping the Numerai leaderboard and return a Pandas DataFrame of the contents.
+        https://numer.ai/tournament
+    """
+    url = 'https://api-tournament.numer.ai/'
+    query = """{ 
+	v2Leaderboard {
+	    returns
+        corrRep
+        fncRep
+        mmcRep
+        username
+        nmrStaked
+        returns 
+        return_52Weeks
+        return_13Weeks
+        rank
+	    }  
+    }"""
+    r = requests.post(url=url, json={'query': query})
+    json_data = json.loads(r.text)
+    df_data = json_data['data']['v2Leaderboard']
+    df = pd.DataFrame(df_data)
+    df.convert_dtypes() # the auto casts the columns into what it thinks is the most appropriate 
+    df['nmrStaked'] = pd.to_numeric(df['nmrStaked'])
+    df = df.rename(columns={'returns': '1D_returns', 'return_13Weeks': '3M_returns', 'return_52Weeks': '1Y_returns'})
+    return df
+
+
