+++ UnfinishedNotebooks/should you day trade or passive fund.txt
@@ -1,0 +1,42 @@
+https://www.ifa.com/articles/despite_brief_reprieve_2018_spiva_report_reveals_active_funds_fail_dent_indexing_lead_-_works/
+
+How many active funds out proform their passive counterparts?
+
+https://faculty.haas.berkeley.edu/odean/papers/Day%20Traders/Day%20Trading%20and%20Learning%20110217.pdf
+2017 Berkly Study on the Day Traders and rationallity. 
+
+From the abstract:
+
+"we document that the aggregate
+performance of day traders is negative, that the vast majority of day traders are
+unprofitable, and many persist despite an extensive experience of losses"
+
+
+"The majority of day traders quit relatively
+quickly (more that 75% of all day traders quit within two years), " (1)
+
+
+"Using Finnish
+data, Seru, Shumway, and Stoffman (2010) document investors not only learn about their
+ability by trading, but also get better with experience. Linnainmaa (2010) calibrates a
+structural model in which investors rationally learn about their ability through trading
+using trading records of active individual Finnish investors. " (1)
+
+"Therefore risk-averse or
+risk-neutral potential speculators with no special prior knowledge of their abilities should
+only “trade to learn” if aggregate speculator profits are positive. In fact, using complete
+data for the Taiwan market, the aggregate performance of day traders net of fees is
+negative in each of the 15 years that we study. A profit-maximizing risk-averse (or riskneutral) Bayesian investor would not enter a market if her expected lifetime profits were
+negative." (2)
+
+
+"The one
+exception to this generalization being Barber, Lee, Liu, and Odean (2010) who identify a
+small subset of day traders (less than 1% of the day trading population) predictably earn
+profits." (6)
+
+
+"In this section,
+we argue that rational learning does not explain behavior of the large population of
+speculative investors for three reasons: aggregate performance is negative, experienced
+speculators lose money, and unprofitable speculators persist."(12)
